+++ src/org/ods/services/OpenShiftService.groovy
@@ -131,6 +131,7 @@ class OpenShiftService {
         def upgradeFlags = defaultFlags.collect { it }
         additionalFlags.collect { upgradeFlags << it }
         valuesFiles.collect { upgradeFlags << "-f ${it}".toString() }
+        values.put('ODS_OPENSHIFT_APP_DOMAIN', getApplicationDomain())
         values.collect { k, v -> upgradeFlags << "--set ${k}=${v}".toString() }
         if (withDiff) {
             def diffFlags = upgradeFlags.findAll { it  }
@@ -148,7 +149,7 @@ class OpenShiftService {
             label: "Upgrade Helm release ${release} in ${project}",
             returnStatus: true
         )
-        if (failed){
+        if (failed) {
             throw new RuntimeException(
                 'Rollout Failed!. ' +
                     "Helm could not install the ${release} in ${project}"
@@ -1316,14 +1317,19 @@ class OpenShiftService {
     }
 
     private void doTailorApply(String project, String tailorParams) {
-        steps.sh(
-            script: """tailor \
-          ${tailorVerboseFlag()} \
-          --non-interactive \
-          -n ${project} \
-          apply ${tailorParams}""",
-            label: "tailor apply for ${project} (${tailorParams})"
-        )
+        try {
+            steps.sh(
+                script: """tailor \
+              ${tailorVerboseFlag()} \
+              --non-interactive \
+              -n ${project} \
+              apply ${tailorParams}""",
+                returnStdout: true,
+                label: "tailor apply for ${project} (${tailorParams})"
+            )
+        } catch (ex) {
+            throw new TailorDeploymentException(ex)
+        }
     }
 
     private void doTailorExport(